move all ffmpeg executor-related health into one place
